refactor: Extract bids component state logic into `BidsComponentState` and consolidate bid row rendering.
